refactor(cdk/drag-drop): move sorting and positioning logic into a separate class
This is an internal reorganization of the drop list to make it easier to eventually add support for mixed orientation. Includes:
* Moving out all the sorting and positioning-related code out of `DropListRef` and into a separate class. The existing logic is very specific to our approach that moves items around using `transform`. The future mixed orientation will likely move the DOM nodes themselves which would've been difficult to manage if we kept the positioning logic within `DropListRef`.
* Moves some of the DOM-related utilities into the `dom` directory in order to reduce clutter in the root of the package.
