+++ src/cdk/drag-drop/dom/client-rect.ts
