+++ src/cdk/drag-drop/dom/clone-node.ts
