+++ src/cdk/drag-drop/dom/transition-duration.ts
