+++ src/cdk/drag-drop/drop-list-ref.ts
@@ -13,17 +13,13 @@ import {ViewportRuler} from '@angular/cdk/scrolling';
 import {_getShadowRoot} from '@angular/cdk/platform';
 import {Subject, Subscription, interval, animationFrameScheduler} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
-import {moveItemInArray} from './drag-utils';
 import {DragDropRegistry} from './drag-drop-registry';
 import {DragRefInternal as DragRef, Point} from './drag-ref';
-import {
-  isPointerNearClientRect,
-  adjustClientRect,
-  getMutableClientRect,
-  isInsideClientRect,
-} from './client-rect';
-import {ParentPositionTracker} from './parent-position-tracker';
-import {combineTransforms, DragCSSStyleDeclaration} from './drag-styling';
+import {isPointerNearClientRect, isInsideClientRect} from './dom/client-rect';
+import {ParentPositionTracker} from './dom/parent-position-tracker';
+import {DragCSSStyleDeclaration} from './dom/styling';
+import {DropListSortStrategy} from './sorting/drop-list-sort-strategy';
+import {SingleAxisSortStrategy} from './sorting/single-axis-sort-strategy';
 
 /**
  * Proximity, as a ratio to width/height, at which a
@@ -36,21 +32,6 @@ const DROP_PROXIMITY_THRESHOLD = 0.05;
  * viewport. The value comes from trying it out manually until it feels right.
  */
 const SCROLL_PROXIMITY_THRESHOLD = 0.05;
-
-/**
- * Entry in the position cache for draggable items.
- * @docs-private
- */
-interface CachedItemPosition {
-  /** Instance of the drag item. */
-  drag: DragRef;
-  /** Dimensions of the item. */
-  clientRect: ClientRect;
-  /** Amount by which the item has been moved since dragging started. */
-  offset: number;
-  /** Inline transform that the drag item had when dragging started. */
-  initialTransform: string;
-}
 
 /** Vertical direction in which we can auto-scroll. */
 const enum AutoScrollVerticalDirection {
@@ -111,7 +92,7 @@ export class DropListRef<T = any> {
    */
   enterPredicate: (drag: DragRef, drop: DropListRef) => boolean = () => true;
 
-  /** Functions that is used to determine whether an item can be sorted into a particular index. */
+  /** Function that is used to determine whether an item can be sorted into a particular index. */
   sortPredicate: (index: number, drag: DragRef, drop: DropListRef) => boolean = () => true;
 
   /** Emits right before dragging has started. */
@@ -155,28 +136,14 @@ export class DropListRef<T = any> {
   /** Whether an item in the list is being dragged. */
   private _isDragging = false;
 
-  /** Cache of the dimensions of all the items inside the container. */
-  private _itemPositions: CachedItemPosition[] = [];
-
   /** Keeps track of the positions of any parent scrollable elements. */
   private _parentPositions: ParentPositionTracker;
 
+  /** Strategy being used to sort items within the list. */
+  private _sortStrategy: DropListSortStrategy<DragRef>;
+
   /** Cached `ClientRect` of the drop list. */
   private _clientRect: ClientRect | undefined;
-
-  /**
-   * Draggable items that are currently active inside the container. Includes the items
-   * from `_draggables`, as well as any items that have been dragged in, but haven't
-   * been dropped yet.
-   */
-  private _activeDraggables: DragRef[];
-
-  /**
-   * Keeps track of the item that was last swapped with the dragged item, as well as what direction
-   * the pointer was moving in when the swap occured and whether the user's pointer continued to
-   * overlap with the swapped item after the swapping occurred.
-   */
-  private _previousSwap = {drag: null as DragRef | null, delta: 0, overlaps: false};
 
   /** Draggable items in the container. */
   private _draggables: readonly DragRef[] = [];
@@ -184,14 +151,8 @@ export class DropListRef<T = any> {
   /** Drop lists that are connected to the current one. */
   private _siblings: readonly DropListRef[] = [];
 
-  /** Direction in which the list is oriented. */
-  private _orientation: 'horizontal' | 'vertical' = 'vertical';
-
   /** Connected siblings that currently have a dragged item. */
   private _activeSiblings = new Set<DropListRef>();
-
-  /** Layout direction of the drop list. */
-  private _direction: Direction = 'ltr';
 
   /** Subscription to the window being scrolled. */
   private _viewportScrollSubscription = Subscription.EMPTY;
@@ -232,6 +193,8 @@ export class DropListRef<T = any> {
     this.withScrollableParents([this.element]);
     _dragDropRegistry.registerDropContainer(this);
     this._parentPositions = new ParentPositionTracker(_document);
+    this._sortStrategy = new SingleAxisSortStrategy(this.element, _dragDropRegistry);
+    this._sortStrategy.withSortPredicate((index, item) => this.sortPredicate(index, item, this));
   }
 
   /** Removes the drop list functionality from the DOM element. */
@@ -262,7 +225,7 @@ export class DropListRef<T = any> {
   }
 
   /**
-   * Emits an event to indicate that the user moved an item into the container.
+   * Attempts to move an item into the container.
    * @param item Item that was moved into the container.
    * @param pointerX Position of the item along the X axis.
    * @param pointerY Position of the item along the Y axis.
@@ -274,65 +237,14 @@ export class DropListRef<T = any> {
 
     // If sorting is disabled, we want the item to return to its starting
     // position if the user is returning it to its initial container.
-    let newIndex: number;
-
-    if (index == null) {
-      newIndex = this.sortingDisabled ? this._draggables.indexOf(item) : -1;
-
-      if (newIndex === -1) {
-        // We use the coordinates of where the item entered the drop
-        // zone to figure out at which index it should be inserted.
-        newIndex = this._getItemIndexFromPointerPosition(item, pointerX, pointerY);
-      }
-    } else {
-      newIndex = index;
+    if (index == null && this.sortingDisabled) {
+      index = this._draggables.indexOf(item);
     }
 
-    const activeDraggables = this._activeDraggables;
-    const currentIndex = activeDraggables.indexOf(item);
-    const placeholder = item.getPlaceholderElement();
-    let newPositionReference: DragRef | undefined = activeDraggables[newIndex];
+    this._sortStrategy.enter(item, pointerX, pointerY, index);
 
-    // If the item at the new position is the same as the item that is being dragged,
-    // it means that we're trying to restore the item to its initial position. In this
-    // case we should use the next item from the list as the reference.
-    if (newPositionReference === item) {
-      newPositionReference = activeDraggables[newIndex + 1];
-    }
-
-    // If we didn't find a new position reference, it means that either the item didn't start off
-    // in this container, or that the item requested to be inserted at the end of the list.
-    if (
-      !newPositionReference &&
-      (newIndex == null || newIndex === -1 || newIndex < activeDraggables.length - 1) &&
-      this._shouldEnterAsFirstChild(pointerX, pointerY)
-    ) {
-      newPositionReference = activeDraggables[0];
-    }
-
-    // Since the item may be in the `activeDraggables` already (e.g. if the user dragged it
-    // into another container and back again), we have to ensure that it isn't duplicated.
-    if (currentIndex > -1) {
-      activeDraggables.splice(currentIndex, 1);
-    }
-
-    // Don't use items that are being dragged as a reference, because
-    // their element has been moved down to the bottom of the body.
-    if (newPositionReference && !this._dragDropRegistry.isDragging(newPositionReference)) {
-      const element = newPositionReference.getRootElement();
-      element.parentElement!.insertBefore(placeholder, element);
-      activeDraggables.splice(newIndex, 0, item);
-    } else {
-      coerceElement(this.element).appendChild(placeholder);
-      activeDraggables.push(item);
-    }
-
-    // The transform needs to be cleared so it doesn't throw off the measurements.
-    placeholder.style.transform = '';
-
-    // Note that the positions were already cached when we called `start` above,
-    // but we need to refresh them since the amount of items has changed and also parent rects.
-    this._cacheItemPositions();
+    // Note that this usually happens inside `_draggingStarted` as well, but the dimensions
+    // can change when the sort strategy moves the item around inside `enter`.
     this._cacheParentPositions();
 
     // Notify siblings at the end so that the item has been inserted into the `activeDraggables`.
@@ -403,7 +315,7 @@ export class DropListRef<T = any> {
       if (draggedItems.every(item => items.indexOf(item) === -1)) {
         this._reset();
       } else {
-        this._cacheItems();
+        this._sortStrategy.withItems(this._draggables);
       }
     }
 
@@ -412,7 +324,7 @@ export class DropListRef<T = any> {
 
   /** Sets the layout direction of the drop list. */
   withDirection(direction: Direction): this {
-    this._direction = direction;
+    this._sortStrategy.direction = direction;
     return this;
   }
 
@@ -431,7 +343,9 @@ export class DropListRef<T = any> {
    * @param orientation New orientation for the container.
    */
   withOrientation(orientation: 'vertical' | 'horizontal'): this {
-    this._orientation = orientation;
+    // TODO(crisbeto): eventually we should be constructing the new sort strategy here based on
+    // the new orientation. For now we can assume that it'll always be `SingleAxisSortStrategy`.
+    (this._sortStrategy as SingleAxisSortStrategy<DragRef>).orientation = orientation;
     return this;
   }
 
@@ -459,19 +373,9 @@ export class DropListRef<T = any> {
    * @param item Item whose index should be determined.
    */
   getItemIndex(item: DragRef): number {
-    if (!this._isDragging) {
-      return this._draggables.indexOf(item);
-    }
-
-    // Items are sorted always by top/left in the cache, however they flow differently in RTL.
-    // The rest of the logic still stands no matter what orientation we're in, however
-    // we need to invert the array when determining the index.
-    const items =
-      this._orientation === 'horizontal' && this._direction === 'rtl'
-        ? this._itemPositions.slice().reverse()
-        : this._itemPositions;
-
-    return items.findIndex(currentItem => currentItem.drag === item);
+    return this._isDragging
+      ? this._sortStrategy.getItemIndex(item)
+      : this._draggables.indexOf(item);
   }
 
   /**
@@ -504,80 +408,16 @@ export class DropListRef<T = any> {
       return;
     }
 
-    const siblings = this._itemPositions;
-    const newIndex = this._getItemIndexFromPointerPosition(item, pointerX, pointerY, pointerDelta);
+    const result = this._sortStrategy.sort(item, pointerX, pointerY, pointerDelta);
 
-    if (newIndex === -1 && siblings.length > 0) {
-      return;
+    if (result) {
+      this.sorted.next({
+        previousIndex: result.previousIndex,
+        currentIndex: result.currentIndex,
+        container: this,
+        item,
+      });
     }
-
-    const isHorizontal = this._orientation === 'horizontal';
-    const currentIndex = siblings.findIndex(currentItem => currentItem.drag === item);
-    const siblingAtNewPosition = siblings[newIndex];
-    const currentPosition = siblings[currentIndex].clientRect;
-    const newPosition = siblingAtNewPosition.clientRect;
-    const delta = currentIndex > newIndex ? 1 : -1;
-
-    // How many pixels the item's placeholder should be offset.
-    const itemOffset = this._getItemOffsetPx(currentPosition, newPosition, delta);
-
-    // How many pixels all the other items should be offset.
-    const siblingOffset = this._getSiblingOffsetPx(currentIndex, siblings, delta);
-
-    // Save the previous order of the items before moving the item to its new index.
-    // We use this to check whether an item has been moved as a result of the sorting.
-    const oldOrder = siblings.slice();
-
-    // Shuffle the array in place.
-    moveItemInArray(siblings, currentIndex, newIndex);
-
-    this.sorted.next({
-      previousIndex: currentIndex,
-      currentIndex: newIndex,
-      container: this,
-      item,
-    });
-
-    siblings.forEach((sibling, index) => {
-      // Don't do anything if the position hasn't changed.
-      if (oldOrder[index] === sibling) {
-        return;
-      }
-
-      const isDraggedItem = sibling.drag === item;
-      const offset = isDraggedItem ? itemOffset : siblingOffset;
-      const elementToOffset = isDraggedItem
-        ? item.getPlaceholderElement()
-        : sibling.drag.getRootElement();
-
-      // Update the offset to reflect the new position.
-      sibling.offset += offset;
-
-      // Since we're moving the items with a `transform`, we need to adjust their cached
-      // client rects to reflect their new position, as well as swap their positions in the cache.
-      // Note that we shouldn't use `getBoundingClientRect` here to update the cache, because the
-      // elements may be mid-animation which will give us a wrong result.
-      if (isHorizontal) {
-        // Round the transforms since some browsers will
-        // blur the elements, for sub-pixel transforms.
-        elementToOffset.style.transform = combineTransforms(
-          `translate3d(${Math.round(sibling.offset)}px, 0, 0)`,
-          sibling.initialTransform,
-        );
-        adjustClientRect(sibling.clientRect, 0, offset);
-      } else {
-        elementToOffset.style.transform = combineTransforms(
-          `translate3d(0, ${Math.round(sibling.offset)}px, 0)`,
-          sibling.initialTransform,
-        );
-        adjustClientRect(sibling.clientRect, offset, 0);
-      }
-    });
-
-    // Note that it's important that we do this after the client rects have been adjusted.
-    this._previousSwap.overlaps = isInsideClientRect(newPosition, pointerX, pointerY);
-    this._previousSwap.drag = siblingAtNewPosition.drag;
-    this._previousSwap.delta = isHorizontal ? pointerDelta.x : pointerDelta.y;
   }
 
   /**
@@ -669,7 +509,8 @@ export class DropListRef<T = any> {
     // that we can't increment/decrement the scroll position.
     this._initialScrollSnap = styles.msScrollSnapType || styles.scrollSnapType || '';
     styles.scrollSnapType = styles.msScrollSnapType = 'none';
-    this._cacheItems();
+    this._sortStrategy.start(this._draggables);
+    this._cacheParentPositions();
     this._viewportScrollSubscription.unsubscribe();
     this._listenToScrollEvents();
   }
@@ -684,27 +525,6 @@ export class DropListRef<T = any> {
     this._clientRect = this._parentPositions.positions.get(element)!.clientRect!;
   }
 
-  /** Refreshes the position cache of the items and sibling containers. */
-  private _cacheItemPositions() {
-    const isHorizontal = this._orientation === 'horizontal';
-
-    this._itemPositions = this._activeDraggables
-      .map(drag => {
-        const elementToMeasure = drag.getVisibleElement();
-        return {
-          drag,
-          offset: 0,
-          initialTransform: elementToMeasure.style.transform || '',
-          clientRect: getMutableClientRect(elementToMeasure),
-        };
-      })
-      .sort((a, b) => {
-        return isHorizontal
-          ? a.clientRect.left - b.clientRect.left
-          : a.clientRect.top - b.clientRect.top;
-      });
-  }
-
   /** Resets the container to its initial state. */
   private _reset() {
     this._isDragging = false;
@@ -712,155 +532,11 @@ export class DropListRef<T = any> {
     const styles = coerceElement(this.element).style as DragCSSStyleDeclaration;
     styles.scrollSnapType = styles.msScrollSnapType = this._initialScrollSnap;
 
-    // TODO(crisbeto): may have to wait for the animations to finish.
-    this._activeDraggables.forEach(item => {
-      const rootElement = item.getRootElement();
-
-      if (rootElement) {
-        const initialTransform = this._itemPositions.find(
-          current => current.drag === item,
-        )?.initialTransform;
-        rootElement.style.transform = initialTransform || '';
-      }
-    });
     this._siblings.forEach(sibling => sibling._stopReceiving(this));
-    this._activeDraggables = [];
-    this._itemPositions = [];
-    this._previousSwap.drag = null;
-    this._previousSwap.delta = 0;
-    this._previousSwap.overlaps = false;
+    this._sortStrategy.reset();
     this._stopScrolling();
     this._viewportScrollSubscription.unsubscribe();
     this._parentPositions.clear();
-  }
-
-  /**
-   * Gets the offset in pixels by which the items that aren't being dragged should be moved.
-   * @param currentIndex Index of the item currently being dragged.
-   * @param siblings All of the items in the list.
-   * @param delta Direction in which the user is moving.
-   */
-  private _getSiblingOffsetPx(currentIndex: number, siblings: CachedItemPosition[], delta: 1 | -1) {
-    const isHorizontal = this._orientation === 'horizontal';
-    const currentPosition = siblings[currentIndex].clientRect;
-    const immediateSibling = siblings[currentIndex + delta * -1];
-    let siblingOffset = currentPosition[isHorizontal ? 'width' : 'height'] * delta;
-
-    if (immediateSibling) {
-      const start = isHorizontal ? 'left' : 'top';
-      const end = isHorizontal ? 'right' : 'bottom';
-
-      // Get the spacing between the start of the current item and the end of the one immediately
-      // after it in the direction in which the user is dragging, or vice versa. We add it to the
-      // offset in order to push the element to where it will be when it's inline and is influenced
-      // by the `margin` of its siblings.
-      if (delta === -1) {
-        siblingOffset -= immediateSibling.clientRect[start] - currentPosition[end];
-      } else {
-        siblingOffset += currentPosition[start] - immediateSibling.clientRect[end];
-      }
-    }
-
-    return siblingOffset;
-  }
-
-  /**
-   * Gets the offset in pixels by which the item that is being dragged should be moved.
-   * @param currentPosition Current position of the item.
-   * @param newPosition Position of the item where the current item should be moved.
-   * @param delta Direction in which the user is moving.
-   */
-  private _getItemOffsetPx(currentPosition: ClientRect, newPosition: ClientRect, delta: 1 | -1) {
-    const isHorizontal = this._orientation === 'horizontal';
-    let itemOffset = isHorizontal
-      ? newPosition.left - currentPosition.left
-      : newPosition.top - currentPosition.top;
-
-    // Account for differences in the item width/height.
-    if (delta === -1) {
-      itemOffset += isHorizontal
-        ? newPosition.width - currentPosition.width
-        : newPosition.height - currentPosition.height;
-    }
-
-    return itemOffset;
-  }
-
-  /**
-   * Checks if pointer is entering in the first position
-   * @param pointerX Position of the user's pointer along the X axis.
-   * @param pointerY Position of the user's pointer along the Y axis.
-   */
-  private _shouldEnterAsFirstChild(pointerX: number, pointerY: number) {
-    if (!this._activeDraggables.length) {
-      return false;
-    }
-
-    const itemPositions = this._itemPositions;
-    const isHorizontal = this._orientation === 'horizontal';
-
-    // `itemPositions` are sorted by position while `activeDraggables` are sorted by child index
-    // check if container is using some sort of "reverse" ordering (eg: flex-direction: row-reverse)
-    const reversed = itemPositions[0].drag !== this._activeDraggables[0];
-    if (reversed) {
-      const lastItemRect = itemPositions[itemPositions.length - 1].clientRect;
-      return isHorizontal ? pointerX >= lastItemRect.right : pointerY >= lastItemRect.bottom;
-    } else {
-      const firstItemRect = itemPositions[0].clientRect;
-      return isHorizontal ? pointerX <= firstItemRect.left : pointerY <= firstItemRect.top;
-    }
-  }
-
-  /**
-   * Gets the index of an item in the drop container, based on the position of the user's pointer.
-   * @param item Item that is being sorted.
-   * @param pointerX Position of the user's pointer along the X axis.
-   * @param pointerY Position of the user's pointer along the Y axis.
-   * @param delta Direction in which the user is moving their pointer.
-   */
-  private _getItemIndexFromPointerPosition(
-    item: DragRef,
-    pointerX: number,
-    pointerY: number,
-    delta?: {x: number; y: number},
-  ): number {
-    const isHorizontal = this._orientation === 'horizontal';
-    const index = this._itemPositions.findIndex(({drag, clientRect}) => {
-      // Skip the item itself.
-      if (drag === item) {
-        return false;
-      }
-
-      if (delta) {
-        const direction = isHorizontal ? delta.x : delta.y;
-
-        // If the user is still hovering over the same item as last time, their cursor hasn't left
-        // the item after we made the swap, and they didn't change the direction in which they're
-        // dragging, we don't consider it a direction swap.
-        if (
-          drag === this._previousSwap.drag &&
-          this._previousSwap.overlaps &&
-          direction === this._previousSwap.delta
-        ) {
-          return false;
-        }
-      }
-
-      return isHorizontal
-        ? // Round these down since most browsers report client rects with
-          // sub-pixel precision, whereas the pointer coordinates are rounded to pixels.
-          pointerX >= Math.floor(clientRect.left) && pointerX < Math.floor(clientRect.right)
-        : pointerY >= Math.floor(clientRect.top) && pointerY < Math.floor(clientRect.bottom);
-    });
-
-    return index === -1 || !this.sortPredicate(index, item, this) ? -1 : index;
-  }
-
-  /** Caches the current items in the list and their positions. */
-  private _cacheItems(): void {
-    this._activeDraggables = this._draggables.slice();
-    this._cacheItemPositions();
-    this._cacheParentPositions();
   }
 
   /** Starts the interval that'll auto-scroll the element. */
@@ -985,23 +661,7 @@ export class DropListRef<T = any> {
           const scrollDifference = this._parentPositions.handleScroll(event);
 
           if (scrollDifference) {
-            // Since we know the amount that the user has scrolled we can shift all of the
-            // client rectangles ourselves. This is cheaper than re-measuring everything and
-            // we can avoid inconsistent behavior where we might be measuring the element before
-            // its position has changed.
-            this._itemPositions.forEach(({clientRect}) => {
-              adjustClientRect(clientRect, scrollDifference.top, scrollDifference.left);
-            });
-
-            // We need two loops for this, because we want all of the cached
-            // positions to be up-to-date before we re-sort the item.
-            this._itemPositions.forEach(({drag}) => {
-              if (this._dragDropRegistry.isDragging(drag)) {
-                // We need to re-sort the item manually, because the pointer move
-                // events won't be dispatched while the user is scrolling.
-                drag._sortFromLastPointerPosition();
-              }
-            });
+            this._sortStrategy.updateOnScroll(scrollDifference.top, scrollDifference.left);
           }
         } else if (this.isReceiving()) {
           this._cacheParentPositions();
@@ -1026,7 +686,9 @@ export class DropListRef<T = any> {
 
   /** Notifies any siblings that may potentially receive the item. */
   private _notifyReceivingSiblings() {
-    const draggedItems = this._activeDraggables.filter(item => item.isDragging());
+    const draggedItems = this._sortStrategy
+      .getActiveItemsSnapshot()
+      .filter(item => item.isDragging());
     this._siblings.forEach(sibling => sibling._startReceiving(this, draggedItems));
   }
 }
